requests: Refactor into protocols module
To allow adding additional protocols in the future, extract the core
protocol functionality from requests.rs to protocols/core.rs.

No functional change intended.
