+++ src/lib.rs
@@ -24,6 +24,7 @@ pub mod io;
 pub mod kernel_launch;
 pub mod locking;
 pub mod mm;
+pub mod protocols;
 pub mod requests;
 pub mod serial;
 pub mod sev;
